requirements.txt
openai-whisper
gTTS
sounddevice
numpy
scipy
pygame
google-generativeai
python-dotenv